Update lookup.js
Blows up when you have both a lookup and tag based multiple relation field on the same form as a tag based lookup.  For example put a multiple relationship of type "lookup" in the user form.  Line 11 blows up as there is no input by that name.  This issue likely affects other areas.  I put this fix into beta2 as a workaround but i'm sure the issue likely happens in beta3 as well.